Added user checks

+++ forum-project/posts/api.py
@@ -8,6 +8,7 @@ api = NinjaAPI(urls_namespace="posts_api")
 
 # Create a post
 @api.post("/posts", response=PostSerializer)
+@api.login_required
 def create_post(request, data: PostSerializer):
     post = Post.objects.create(**data.dict())
     return post
@@ -27,8 +28,11 @@ def get_post(request, post_id: int):
 
 # update a post
 @api.put("/posts/{post_id}", response=PostSerializer)
+@api.login_required
 def update_post(request, post_id: int, data: PostSerializer):
     post = Post.objects.get(id=post_id)
+    if post.user != request.user:
+        return 403, {"message": "You do not have permission to update this post."}
     for attr, value in data.dict().items():
         setattr(post, attr, value)
     post.save()
@@ -36,9 +40,12 @@ def update_post(request, post_id: int, data: PostSerializer):
 
 # delete a post
 @api.delete("/posts/{post_id}", response={200: str, 404: str})
+@api.login_required
 def delete_post(request, post_id: int):
     try:
         post = Post.objects.get(id=post_id)
+        if post.user != request.user:
+            return 403, {"message": "You do not have permission to delete this post."}
         post.delete()
         return 200, {"message": "Post deleted successfully"}
     except Post.DoesNotExist:
